+++ test/fixtures/database.sql
@@ -1,3 +1,3 @@
-CREATE EXTENSION IF NOT EXISTS pgcrypto;
-
-ALTER DATABASE :db SET request.jwt.claim.id = '-1';
+set client_min_messages to warning;
+DROP SCHEMA IF EXISTS test, private, postgrest, jwt, تست CASCADE;
+DROP TYPE IF EXISTS jwt_token CASCADE;